requirements.txt
blackboxprotobuf=1.0.1
frida=16.1.4
frida-tools=12.3.0
fpdf=1.7.2
pyyaml=6.0.1